Move get_kernel_desc where it is used, some cleanups
