+++ pyproject.toml
@@ -25,7 +25,7 @@ write_template = "__version__ = '{}'"  # optional, default to "{}"
 requires = ["pdm-pep517>=0.12.0"]
 build-backend = "pdm.pep517.api"
 
-# Ruff is installed using pre-commit
+# Ruff is installed and run using pre-commit
 [tool.ruff]
 line-length = 88
 target-version = "py38"
